refactor(3000): Clean up sidebar files structure
