+++ frontend/src/layout/navigation-3000/components/Sidebar.tsx
@@ -5,7 +5,8 @@ import { IconClose, IconMagnifier } from 'lib/lemon-ui/icons'
 import React, { useRef, useState } from 'react'
 import { navigation3000Logic } from '../navigationLogic'
 import { KeyboardShortcut } from './KeyboardShortcut'
-import { SidebarAccordion, SidebarList } from './SidebarAccordion'
+import { SidebarAccordion } from './SidebarAccordion'
+import { SidebarList } from './SidebarList'
 import { Accordion, BasicListItem, ExtendedListItem, SidebarLogic } from '../types'
 import { Spinner, SpinnerOverlay } from 'lib/lemon-ui/Spinner'
 import { useDebouncedCallback } from 'use-debounce'
