+++ frontend/src/layout/navigation-3000/sidebars/cohorts.ts
@@ -3,7 +3,7 @@ import { sceneLogic } from 'scenes/sceneLogic'
 import { Scene } from 'scenes/sceneTypes'
 import { urls } from 'scenes/urls'
 import { ExtendedListItem } from '../types'
-import type { cohortsSidebarLogicType } from './cohortsSidebarLogicType'
+import type { cohortsSidebarLogicType } from './cohortsType'
 import Fuse from 'fuse.js'
 import { CohortType } from '~/types'
 import { subscriptions } from 'kea-subscriptions'
