+++ frontend/src/layout/navigation-3000/sidebars/dashboards.ts
@@ -4,7 +4,7 @@ import { Scene } from 'scenes/sceneTypes'
 import { urls } from 'scenes/urls'
 import { dashboardsModel } from '~/models/dashboardsModel'
 import { ExtendedListItem } from '../types'
-import type { dashboardsSidebarLogicType } from './dashboardsSidebarLogicType'
+import type { dashboardsSidebarLogicType } from './dashboardsType'
 import Fuse from 'fuse.js'
 import { DashboardType } from '~/types'
 import { subscriptions } from 'kea-subscriptions'
