+++ frontend/src/layout/navigation-3000/sidebars/featureFlags.tsx
@@ -6,7 +6,7 @@ import { sceneLogic } from 'scenes/sceneLogic'
 import { Scene } from 'scenes/sceneTypes'
 import { urls } from 'scenes/urls'
 import { ExtendedListItem } from '../types'
-import type { featureFlagsSidebarLogicType } from './featureFlagsSidebarLogicType'
+import type { featureFlagsSidebarLogicType } from './featureFlagsType'
 import Fuse from 'fuse.js'
 import { FeatureFlagType } from '~/types'
 import { subscriptions } from 'kea-subscriptions'
